Add validity of predicate translations check and support "zero length" predicates
Avoid to generate conditions that are indeed wrong because of the mismatch
between the subject type and the predicate translation. For example, if the
subject is of type boolean, a translation of the predicate like "==0" would
be considered wrong and discarded. At the moment, validity check of predicate
translations is implemented only for subjects that are method's arguments.

Also, predicates like "has zero length" on arrays are now correctly supported.

This change improves precision and recall.

Resolves: #65, #66, #67